+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added kotsu exception DeprecatedEntityError, and use instead of ValueError for deprecated entity
+
 ### Changed
 - Replaced `entry_point=None` in entity spec with `deprecated=True`
 - Changed passing `artefacts_store_dir` to passing `validation_artefacts_dir` and